Move physics, transport and math to theory subpackage
