+++ plasmapy/examples/plot_physics.py
@@ -6,10 +6,9 @@ Let's try to look at ITER plasma conditions using the `physics` subpackage.
 """
 
 from astropy import units as u
-from plasmapy import physics
+from plasmapy.theory import physics
 import matplotlib.pyplot as plt
 import numpy as np
-from mpl_toolkits.mplot3d import Axes3D
 
 ######################################################
 # The radius of electric field shielding clouds, also known as the Debye length,
