+++ plasmapy/theory/mathematics/mathematics.py
