+++ plasmapy/theory/mathematics/tests/test_dispersion.py
