+++ plasmapy/theory/physics/tests/test_distribution.py
@@ -5,7 +5,7 @@ import pytest
 from astropy import units as u
 import scipy.integrate as spint
 
-from ...constants import (m_p, m_e, c, mu0, k_B, e, eps0, pi)
+from plasmapy.constants import (m_p, m_e, c, mu0, k_B, e, eps0, pi)
 from ..distribution import (Maxwellian_1D,
                             Maxwellian_speed_1D,
                             Maxwellian_velocity_2D,
