+++ plasmapy/theory/physics/tests/test_parameters_cython.py
@@ -2,16 +2,10 @@
 
 import numpy as np
 import pytest
-from astropy import units as u
-from warnings import simplefilter
 
-from plasmapy.utils.exceptions import RelativityWarning, RelativityError
-from plasmapy.utils.exceptions import PhysicsError
-from plasmapy.constants import c, m_p, m_e, e, mu0
-
-import plasmapy.physics.parameters_cython
-from plasmapy.physics.parameters_cython import (thermal_speed,
-                                 )
+import plasmapy.theory.physics.parameters_cython
+from plasmapy.theory.physics import (thermal_speed,
+                                     )
 def test_thermal_speed():
     r"""Test for cythonized version of thermal_speed()."""
     trueVal = 593083.619464999
