+++ plasmapy/theory/physics/tests/test_quantum.py
@@ -1,8 +1,8 @@
 import numpy as np
 import pytest
 import astropy.units as u
-from ...constants import c, h
-from ...utils.exceptions import RelativityError
+from plasmapy.constants import c, h
+from plasmapy.utils.exceptions import RelativityError
 from ..quantum import (deBroglie_wavelength,
                        thermal_deBroglie_wavelength,
                        Fermi_energy,
