+++ plasmapy/theory/physics/tests/test_relativity.py
@@ -3,9 +3,9 @@
 import pytest
 import numpy as np
 from astropy import units as u
-from ...constants import c
+from plasmapy.constants import c
 from ..relativity import Lorentz_factor
-from ...utils.exceptions import RelativityError
+from plasmapy.utils.exceptions import RelativityError
 
 
 def test_Lorentz_factor():
