+++ plasmapy/theory/transport/tests/test_collisions.py
@@ -3,20 +3,20 @@ import pytest
 from astropy import units as u
 from astropy.tests.helper import assert_quantity_allclose
 
-from plasmapy.transport import (Coulomb_logarithm,
-                                impact_parameter_perp,
-                                impact_parameter,
-                                collision_frequency,
-                                mean_free_path,
-                                mobility,
-                                Knudsen_number,
-                                coupling_parameter)
-from plasmapy.transport import (Spitzer_resistivity,
-                                fundamental_electron_collision_freq,
-                                fundamental_ion_collision_freq)
+from plasmapy.theory.transport import (Coulomb_logarithm,
+                                       impact_parameter_perp,
+                                       impact_parameter,
+                                       collision_frequency,
+                                       mean_free_path,
+                                       mobility,
+                                       Knudsen_number,
+                                       coupling_parameter)
+from plasmapy.theory.transport import (Spitzer_resistivity,
+                                       fundamental_electron_collision_freq,
+                                       fundamental_ion_collision_freq)
 from plasmapy.utils import exceptions
 from plasmapy.utils.pytest_helpers import assert_can_handle_nparray
-from plasmapy.constants import m_p, m_e, c
+from plasmapy.constants import c
 
 
 class Test_Coulomb_logarithm:
